demand: Move the batch routing demand types to transition
These types are not related to trRouting and should not be part of the
API. They are moved to the transitDemand module of transition-common.

TransitBatchCalculationResult is moved to the batchCalculation module of
transition-common.
