+++ packages/chaire-lib-common/src/api/TrRouting/base.ts
@@ -4,48 +4,6 @@
  * This file is licensed under the MIT License.
  * License text available at https://opensource.org/licenses/MIT
  */
-import { ErrorMessage } from '../../utils/TrError';
-
-// TODO These batch demand attributes are not trRouting specific. They should go elsewhere
-export type TransitBatchDemandFromCsvAttributes = {
-    calculationName: string;
-    projection: string;
-    idAttribute: string;
-    timeAttributeDepartureOrArrival: 'arrival' | 'departure';
-    timeFormat: string;
-    timeAttribute: string;
-    withGeometries: boolean;
-    detailed: boolean;
-    cpuCount: number;
-    csvFile: { location: 'upload'; filename: string } | { location: 'server'; fromJob: number };
-};
-
-export type TransitBatchRoutingDemandFromCsvAttributes = TransitBatchDemandFromCsvAttributes & {
-    originXAttribute: string;
-    originYAttribute: string;
-    destinationXAttribute: string;
-    destinationYAttribute: string;
-    saveToDb: false | { type: 'new'; dataSourceName: string } | { type: 'overwrite'; dataSourceId: string };
-};
-
-export type TransitBatchRoutingDemandAttributes = {
-    type: 'csv';
-    configuration: TransitBatchRoutingDemandFromCsvAttributes;
-};
-
-export type TransitBatchAccessibilityMapAttributes = TransitBatchDemandFromCsvAttributes & {
-    xAttribute: string;
-    yAttribute: string;
-};
-
-export interface TransitBatchCalculationResult {
-    calculationName: string;
-    detailed: boolean;
-    completed: boolean;
-    warnings: ErrorMessage[];
-    errors: ErrorMessage[];
-}
-
 export interface TransitRouteQueryOptions {
     /**
      * An array containing the origin and destination points.
