+++ packages/transition-frontend/src/components/forms/batchCalculation/BatchCalculationList.tsx
@@ -14,7 +14,7 @@ import Button from 'chaire-lib-frontend/lib/components/input/Button';
 import ExecutableJobComponent from '../../parts/executableJob/ExecutableJobComponent';
 import TransitBatchRoutingCalculator from 'transition-common/lib/services/transitRouting/TransitBatchRoutingCalculator';
 import { BatchCalculationParameters } from 'transition-common/lib/services/batchCalculation/types';
-import { TransitBatchRoutingDemandAttributes } from 'chaire-lib-common/lib/api/TrRouting';
+import { TransitBatchRoutingDemandAttributes } from 'transition-common/lib/services/transitDemand/types';
 import FormErrors from 'chaire-lib-frontend/lib/components/pageParts/FormErrors';
 import TrError, { ErrorMessage } from 'chaire-lib-common/lib/utils/TrError';
