+++ packages/transition-frontend/src/services/accessibilityMap/BatchAccessibilityMapCalculator.ts
@@ -7,10 +7,8 @@
 import TrError from 'chaire-lib-common/lib/utils/TrError';
 import serviceLocator from 'chaire-lib-common/lib/utils/ServiceLocator';
 import TransitBatchAccessibilityMap from 'transition-common/lib/services/accessibilityMap/TransitBatchAccessibilityMap';
-import {
-    TransitBatchAccessibilityMapAttributes as TransitBatchAccessibilityMapAttributesBase,
-    TransitBatchCalculationResult
-} from 'chaire-lib-common/lib/api/TrRouting';
+import { TransitBatchAccessibilityMapAttributes as TransitBatchAccessibilityMapAttributesBase } from 'transition-common/lib/services/transitDemand/types';
+import { TransitBatchCalculationResult } from 'transition-common/lib/services/batchCalculation/types';
 import { TrRoutingConstants } from 'chaire-lib-common/lib/api/TrRouting';
 import AccessibilityMapRouting from 'transition-common/lib/services/accessibilityMap/TransitAccessibilityMapRouting';
 import * as Status from 'chaire-lib-common/lib/utils/Status';
